backend/requirements.txt
# Render’s Python deploy flow expects a build step like pip install -r requirements.txt.

fastapi
uvicorn[standard]
python-dotenv
requests
playwright
pyyaml
openai
weaviate-client
pydantic
psycopg[binary]